+++ playground/core/SoundObject.hpp
@@ -32,12 +32,12 @@ public:
     void ProcessBuffer(float *inData, float *outData, int numFrames, double curTime);
     
 protected:
-    int state;              /* 0 - stopped  1 - playing */
+    uint8_t state;              /* 0 - stopped  1 - playing */
     
     static constexpr int TABLE_SIZE = 2048;
     
     float sine[TABLE_SIZE];
-    int leftPhase;
-    int rightPhase;
+    uint32_t leftPhase;
+    uint32_t rightPhase;
     double lastTime;
 };
